added skip detection

+++ R/log.R
@@ -3,5 +3,5 @@ saveLog <- function(e)UseMethod("saveLog")
 saveLog.default <- function(e){
   # save log
   suppressMessages(suppressWarnings(
-    saveRDS(e$log, file.path(e$udat, paste0(as.numeric(Sys.time()), ".swlog")))))
+    saveRDS(e$log, file.path(e$udat, paste0(as.integer(Sys.time()), ".swlog")))))
 }
